I added tests for returning bottom left and right corners as closest points.
Caught bug in logic which resulted in correct answer for positive y, but incorrect for negative values. The closing bracket was in the wrong place,
so the comparison was actually happening between the difference y element and half the box height, the absolute value of which would then be taken.
A negative y value would be less than zero, the absolute of which would be zero and thus the if statement wouldn't be entered and the offset
Vector would be set to (0, 0);